+++ src/bacnet/basic/service/h_apdu.c
@@ -88,7 +88,7 @@ static BACNET_SERVICES_SUPPORTED
  * @brief get the local network priority
  * @return local network priority
  */
-uint16_t apdu_network_priority(void)
+uint8_t apdu_network_priority(void)
 {
     return Local_Network_Priority;
 }
@@ -97,7 +97,7 @@ uint16_t apdu_network_priority(void)
  * @brief set the local network priority
  * @param net - local network priority
  */
-void apdu_network_priority_set(uint16_t pri)
+void apdu_network_priority_set(uint8_t pri)
 {
     Local_Network_Priority = pri & 0x03;
 }
